Refactor install.sh script to remove i386 architecture and ensure Ubuntu 22.04 compatibility

+++ install.sh
@@ -107,16 +107,18 @@ sudo chown root:root /etc/apt/preferences.d/no-snap.pref
 judge "Remove snap"
 
 # Docker source
-print_ok "Setting docker..."
-sudo install -m 0755 -d /etc/apt/keyrings
-sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
-sudo chmod a+r /etc/apt/keyrings/docker.asc
-judge "Trust docker gpg"
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
-  $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
-  sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-judge "Setting docker"
+if [ "$IN_CONTAINER" != "TRUE" ]; then
+  print_ok "Setting docker..."
+  sudo install -m 0755 -d /etc/apt/keyrings
+  sudo curl -fsSL https://download.docker.com/linux/ubuntu/gpg -o /etc/apt/keyrings/docker.asc
+  sudo chmod a+r /etc/apt/keyrings/docker.asc
+  judge "Trust docker gpg"
+  echo \
+    "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
+    $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
+    sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+  judge "Setting docker"
+fi
 
 # Google Chrome Source
 print_ok "Setting google chrome..."
@@ -201,7 +203,6 @@ sudo apt install -y \
   openjdk-17-jdk default-jre\
   ruby\
   python3-apt python3-pip python-is-python3\
-  docker-ce docker-ce-cli containerd.io docker-compose\
   hugo\
   spotify-client\
   adb\
@@ -222,6 +223,12 @@ sudo apt install -y \
   aisleriot\
   qtwayland5
 judge "Install apt softwares"
+
+if [ "$IN_CONTAINER" != "TRUE" ]; then
+  print_ok "Installing docker..."
+  sudo apt install -y docker-ce docker-ce-cli containerd.io docker-compose\
+  judge "Install docker"
+fi
 
 # WeChat
 print_ok "Setting wechat..."
@@ -247,9 +254,11 @@ sudo apt autoremove -y gnome-terminal > /dev/null 2>&1 || true
 judge "Remove obsolete gnome apps"
 
 # Add current user as docker.
-print_ok "Adding $USER to docker group..."
-sudo gpasswd -a $USER docker
-judge "Add $USER to docker group"
+if [ "$IN_CONTAINER" != "TRUE" ]; then
+  print_ok "Adding $USER to docker group..."
+  sudo gpasswd -a $USER docker
+  judge "Add $USER to docker group"
+fi
 
 # NPM
 print_ok "Installing npm global packages..."
@@ -292,18 +301,20 @@ else
 fi
 
 # Installing docker-desktop
-if ! dpkg -s docker-desktop > /dev/null 2>&1; then
-    print_ok "docker-desktop is not installed, downloading and installing..."
-    # Download the deb package
-    wget https://desktop.docker.com/linux/main/amd64/docker-desktop-4.28.0-amd64.deb
-    # Install the package
-    sudo dpkg -i docker-desktop-4.28.0-amd64.deb
-    sudo apt install --fix-broken -y
-    judge "Install docker-desktop"
-    # Remove the package file
-    rm docker-desktop-4.28.0-amd64.deb
-else
-    print_ok "docker-desktop is already installed"
+if [ "$IN_CONTAINER" != "TRUE" ]; then
+  if ! dpkg -s docker-desktop > /dev/null 2>&1; then
+      print_ok "docker-desktop is not installed, downloading and installing..."
+      # Download the deb package
+      wget https://desktop.docker.com/linux/main/amd64/docker-desktop-4.28.0-amd64.deb
+      # Install the package
+      sudo dpkg -i docker-desktop-4.28.0-amd64.deb
+      sudo apt install --fix-broken -y
+      judge "Install docker-desktop"
+      # Remove the package file
+      rm docker-desktop-4.28.0-amd64.deb
+  else
+      print_ok "docker-desktop is already installed"
+  fi
 fi
 
 # Chinese input
